fix: run xlsx conversion in worker thread
